Make view Generator smaller
