+++ src/types/GenericStringObject.ts
@@ -1,0 +1,3 @@
+type GenericStringObject = { [key: string]: string };
+
+export default GenericStringObject;
