+++ src/types/SuggestionList.ts
@@ -1,7 +1,9 @@
+import GenericStringObject from './GenericStringObject';
+
 type SuggestionList = {
   categoryTitle: string;
   categoryContentType: string;
-  suggestions: { [key: string]: string[] };
+  suggestions: GenericStringObject[];
 };
 
 export default SuggestionList;
