--- src/types/SuggestionTextProps.ts
@@ -1,3 +1,0 @@
-type SuggestionTextProps = { [key: string]: string };
-
-export default SuggestionTextProps;
