+++ src/views/Generator/index.css
@@ -11,7 +11,7 @@
   grid-template-columns: 1fr;
   width: 100%;
   min-width: 180px;
-  min-height: 220px;
+  min-height: 275px;
   padding: 2em 0.5em;
   border-radius: var(--border-radius);
   margin: 0 0 2em 0;
